Rename app cast component folder
